back/requirements.txt
python-dotenv~=1.0.0
requests~=2.32.1
Pillow~=10.4.0
ollama~=0.4.2
python-json-logger~=2.0.7